pyproject.toml
[project]
name = "pat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Peter Hoburg", email = "peterHoburg@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "fastapi[standard]>=0.115.12",
    "psycopg[binary,pool]>=3.2.6",
    "pydantic-ai[logfire]>=0.1.4",
    "pydantic-settings>=2.9.1",
    "sqlalchemy[asyncio]>=2.0.40",
    "tenacity>=9.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "inline-snapshot>=0.22.3",
    "pyright>=1.1.400",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.6",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 120
indent-width = 4
target-version = "py312"
fix = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100",
    "D101",
    "D103",
    "D104",
    "D105",
    "D400",
    "D401",
    "D415",

    "D203",
    "D213",

    "COM812",
]

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "D",
    "ANN001",
    "ANN201",
    "S101",
    "T201",
]
"alembic/**" = [
    "INP001",
    "ERA001",
]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
asyncio_mode = "auto"
